+++ system/modules/core/elements/ContentElement.php
@@ -133,6 +133,16 @@ abstract class ContentElement extends \Frontend
 
 
 	/**
+	 * Return the model
+	 * @return \Model
+	 */
+	public function getModel()
+	{
+		return $this->objModel;
+	}
+
+
+	/**
 	 * Parse the template
 	 * @return string
 	 */