+++ system/modules/core/modules/ModuleArticle.php
@@ -330,7 +330,6 @@ class ModuleArticle extends \Module
 
 		// Include library
 		require_once TL_ROOT . '/system/config/tcpdf.php';
-		require_once TL_ROOT . '/system/modules/core/vendor/tcpdf/tcpdf.php';
 
 		// Create new PDF document
 		$pdf = new \TCPDF(PDF_PAGE_ORIENTATION, PDF_UNIT, PDF_PAGE_FORMAT, true);