+++ system/docs/CHANGELOG.md
@@ -4,5 +4,8 @@ Contao Open Source CMS changelog
 Version 3.3.beta1 (2014-XX-XX)
 ------------------------------
 
+### Improved
+Support the "HAVING" command in the `Model\QueryBuilder` class (see #6446).
+
 ### Changed
 Use class constants for `BackendUser::isAllowed()`.