+++ system/docs/CHANGELOG.md
@@ -5,6 +5,9 @@ Version 3.3.beta1 (2014-XX-XX)
 ------------------------------
 
 ### Improved
+Add a `getModel()` method to modules, elements and hybrids (see #6492).
+
+### Improved
 Support the "HAVING" command in the `Model\QueryBuilder` class (see #6446).
 
 ### Changed