+++ system/docs/CHANGELOG.md
@@ -4,6 +4,10 @@ Contao Open Source CMS changelog
 Version 3.3.beta1 (2014-XX-XX)
 ------------------------------
 
+### New
+Added a `DcaLoader` class to decouple the DCA loading process (see #5441). DCAs
+can now be loaded anywhere using `DataContainer::load()`.
+
 ### Changed
 Convert slashes to hyphens in the `standardize()` function (see #6396).