+++ system/docs/CHANGELOG.md
@@ -4,6 +4,9 @@ Contao Open Source CMS changelog
 Version 3.3.beta1 (2014-XX-XX)
 ------------------------------
 
+### New
+Support editing the front end preview page via the "url" parameter (see #6471).
+
 ### Improved
 Do not combine .js and .css files when running in debug mode (see #6450).